+++ Packages/com.makaria122.re-world-maker/Authoring/REAudioAction.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 namespace REWorldMaker
 {
-    [AddComponentMenu("RE World Maker/Actions/REW Audio Action"), DisallowMultipleComponent, RequireComponent(typeof(AudioSource))]
+    [AddComponentMenu("RE World Maker/Actions/REW Action Audio"), DisallowMultipleComponent, RequireComponent(typeof(AudioSource))]
     public sealed class REAudioAction : REAction
     { public AudioSource audioSource; public REAudioOperation operation; private void Reset() { audioSource = GetComponent<AudioSource>(); } }
 }