+++ Packages/com.makaria122.re-world-maker/Authoring/REInteractTrigger.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 namespace REWorldMaker
 {
-    [AddComponentMenu("RE World Maker/Triggers/REW Interact Trigger"), DisallowMultipleComponent]
+    [AddComponentMenu("RE World Maker/Triggers/REW Trigger Interact"), DisallowMultipleComponent]
     public sealed class REInteractTrigger : RETrigger
     {
         public string interactionText = "Use";