+++ Packages/com.makaria122.re-world-maker/Authoring/REPickupUseTrigger.cs
@@ -3,7 +3,7 @@ using VRC.SDK3.Components;
 
 namespace REWorldMaker
 {
-    [AddComponentMenu("RE World Maker/Triggers/REW Pickup Use Trigger")]
+    [AddComponentMenu("RE World Maker/Triggers/REW Trigger Pickup Use")]
     [DisallowMultipleComponent]
     [RequireComponent(typeof(BoxCollider), typeof(VRCPickup))]
     public sealed class REPickupUseTrigger : RETrigger