+++ Packages/com.makaria122.re-world-maker/Authoring/RETeleportAction.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 namespace REWorldMaker
 {
-    [AddComponentMenu("RE World Maker/Actions/REW Teleport Action"), DisallowMultipleComponent]
+    [AddComponentMenu("RE World Maker/Actions/REW Action Teleport"), DisallowMultipleComponent]
     public sealed class RETeleportAction : REAction { public Transform destination; }
 }