+++ Packages/com.makaria122.re-world-maker/Authoring/REToggleObjectAction.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 namespace REWorldMaker
 {
-    [AddComponentMenu("RE World Maker/Actions/REW Toggle Object Action"), DisallowMultipleComponent]
+    [AddComponentMenu("RE World Maker/Actions/REW Action Toggle Object"), DisallowMultipleComponent]
     public sealed class REToggleObjectAction : REAction { public GameObject[] targetObjects; public REToggleOperation operation; }
 }