+++ Packages/com.makaria122.re-world-maker/Authoring/REWorldStartTrigger.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 namespace REWorldMaker
 {
-    [AddComponentMenu("RE World Maker/Triggers/REW World Start Trigger"), DisallowMultipleComponent]
+    [AddComponentMenu("RE World Maker/Triggers/REW Trigger World Start"), DisallowMultipleComponent]
     public sealed class REWorldStartTrigger : RETrigger { [Min(0f)] public float delaySeconds; }
 }